reformat imports and class decl for FrameVar and ScratchVar
