+++ pyteal/ast/__init__.py
@@ -155,7 +155,7 @@ from pyteal.ast.while_ import While
 from pyteal.ast.for_ import For
 from pyteal.ast.break_ import Break
 from pyteal.ast.continue_ import Continue
-from pyteal.ast.frame import Proto, FrameDig, FrameBury, PopN, DupN, Bury
+from pyteal.ast.frame import Proto, FrameDig, FrameBury, PopN, DupN, Bury, FrameStorage
 
 # misc
 from pyteal.ast.scratch import (
@@ -344,6 +344,7 @@ __all__ = [
     "Bury",
     "DupN",
     "PopN",
+    "FrameStorage",
     "Router",
     "CallConfig",
     "MethodConfig",
